backend/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt